backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
rdkit-pypi
requests
python-dotenv
emails
jinja2
pandas
openpyxl
aiofiles
python-barcode
google-auth>=2.0.0